Refactor Assembly class. Eliminate `routes` as a public API concern
